Move metrics logic to its own package
